fix [bugs:#1610] SysFileSearch Crashes; add tests; update rexxref
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/main/trunk@11847 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ interpreter/runtime/Interpreter.cpp
@@ -753,7 +753,7 @@ const char* REXXENTRY RexxGetErrorMessageByNumber(int msgid)
  */
 RexxString* Interpreter::getMessageText(wholenumber_t code)
 {
-    const char *message = RexxGetErrorMessage(code);
+    const char *message = RexxGetErrorMessage((int)code);
     if (message != NULL)
     {
         return new_string(message);
